RAC-260: add forward ref unit test
