+++ akeneo-design-system/src/components/Button/Button.tsx
@@ -162,4 +162,3 @@ const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
 );
 
 export {Button};
-export type {ButtonSize, Level as ButtonLevel};
